refactor: address PR review feedback for JWT decoder
Code organization improvements:
- Extract JWT utility functions to app/utils/jwt.ts
- Add comprehensive JSDoc comments (80%+ coverage)
- Follow established patterns from other utility modules

UX improvements:
- Remove redundant alert() dialogs
- Use inline error messages exclusively
- Better accessibility for screen readers

Styling improvements:
- Replace inline styles with CSS classes
- Add JWT-specific classes (.jwt-output-section, .jwt-copy-button, etc.)
- Use Material Design 3 color system for error messages

Test improvements:
- Update E2E tests to check error messages instead of alert dialogs
- Maintain comprehensive test coverage

All unit tests and build pass successfully.
